Added a script to compare cdm expected output with transform output. It handles different person_id integer labelling, missing data entered as 0 (where it applied to the whole field) and removes times for comparisons (mainly to remove 00:00:00.00000, but actually removes any time).
It does not remove all 0s from any column/field, as it is possible to have a valid measurement of 0.
It does not check for a time of 00:00:00.00000, it just removes all times.
It does not check the dimensions are equal - it may be possible that additional columns, especially empty ones, may have no impact.