Update metrics.py
Created dataclasses to use instead of the key building code in run.

The DataKey class makes keys for instances of CountData, so you can add them to Metric.datasummary

A MapstreamSummaryRow encapsulates the logic for a row in Metric.get_mapstream_summary, so that get_mapstream_summary can build the string from a list of rows and get the header from the class. You can make the TSV string for it, or use it as a dictionary if you want to pass it about.

The `increment_key_counts` function in run.py was very confusing, so I've replaced it with a Metrics method, increment_with_datacol, which performs the same logic, but is bound to the class rather than floating around, and applies the same logic, without the hard to read strings.