Hashable Mappings?
I've added a hashable MappingKey to the mappingrules so that the fast lookup can be maintained. I've left the other properties on the Mapping class in case they're needed. If not, then a Dict[MappingKey, term_mapping] would be better.

Eventually, bringing this together with the DataKey from metrics will be the way to go I suspect